requirements.txt
pyserial>=4.0
rich>=4.0